telegram_bot/requirements.txt
aiogram>=3.0
openai
httpx
python-dotenv
aiohttp-socks
